Stats: Reduce graph creation complexity.
Stop searching for contributions using a regex containing a username.
Instead, make a dictionary with the username as a key upfront. The
dictionary contains a sorted list of contribution dates.

Additionally, save the line split by tabs in a variable, rather than
splitting it multiple times.

This reduces the time searching for rust issue responders from 8 minutes
to 2 seconds. :D

Signed-off-by: Sarah Sharp <sharp@otter.technology>