Statistics: Hypothesis testing example.
In order to determine whether sentiment has an impact on participation,
I need to use hypothesis testing. ghrusthighfive.py is an example of
hypothesis testing that I can do without wrestling with the sentiment
analysis library.

rust-highfive is a github bot that suggests one person to review a pull
request, and notifies them with an @username. Hypothesis testing helps
us answer the question:

"Are pull requests that rust-highfive commented on more likely to be
merged?"

The answer is yes, with 99% confidence.

80.5% of the pull requests that rust-highfive commented on were merged.
58.8% of the pull requests that rust-highfive didn't comment on were
merged. With a sample size of 7532 and 10932, respectively, we can be
very confident of our results. I verified the math with this handy
calculator from Evan Miller:

http://www.evanmiller.org/ab-testing/t-test.html#!0.804965/0.396227/7532;0.588181/0.492163/10932@97

Signed-off-by: Sarah Sharp <sharp@otter.technology>