Hypothesis: Make output clearer.
Add a --debug flag to hide the detailed statistics that need to be
plugged into Evan Miller's site to double check my math.

Add additional newlines between the two tests we're running to make it
easier to parse the statistics.

Move the statistics differences to the front of the sentences to make it
easier to visually parse. Move the difference between the two means to
the middle of the sentences to break up the monotony of to many numbers.

Signed-off-by: Sarah Sharp <sharp@otter.technology>