rust-highfive: Measure 1 year before and after.
Do this for both bors and rust-fivehigh testing. The average amount of
time a merged pull request is open has been increasing over time, so we
need to bound the measurements to one year before and after.

Since this is a upwards trend we could use smoothing techniques to find
a trend line, and subtract that trend in order to get the "noisy" data.
Probably not something we can work on soon, and it won't hold true for
all data.

Note that the test for whether more pull requests were merged uses
two-means hypothesis testing, where it should use a Chi squared test. I
couldn't figure out how to reproduce the math for a Chi squared test, so
I've been plugging the count and successful pull request numbers into
Evan Miller's website.

Signed-off-by: Sarah Sharp <sharp@otter.technology>