requirements.txt: Return statistics
ghstats.py requires statistics in order to run. This requirement was
originally featured in requirements.txt, but was removed by commit
3a79f8. That removal seems to be a mistake, as requiring statistics
do not harm the functionality of ghrusthighfive.py that commit
introduced, and it's clearly needed by ghstats.py.

This patch returns the statistics requirement in order to unbreak
ghstats.py.

Note that python 3.4 and above already ship the statistics module,
so this requirement is only needed for older versions (notable, the
entire python 2 generation).

This commit solves issue #31.