Language: Break out curse words we haven't found yet.
There are several curse words that we haven't found in the github repos.
That's a good thing, however, it means we don't have a real sentence to
feed into the sentiment model that contains that curse word. Having a
one-word sentence in the training model doesn't help the neural net
learn how to propagate the negative sentiment up the sentence. I found a
couple of the words on the debian-devel mailing list, but soon stopped
looking because it was too time consuming.

So, for now, break out the unfound curse words into their own file.

Signed-off-by: Sarah Sharp <sharp@otter.technology>