Word hypothesis: Reviewers don't need to be sorted.
Since the reviewer dictionary's key is the path to the comment json
file, we'll have one unique entry per file. That means we don't need to
sort the list by date before turning it into a dict. We also don't need
to create a list for each dict entry, but for now just turn it into a
list with one element to keep the code simpler.

Signed-off-by: Sarah Sharp <sharp@otter.technology>