Sentiment: emoji list.
I don't want to feed the raw emoji into the sentiment training, because
the parser doesn't group it the same when it's used in a sentence.
Instead, we need to find real-world examples in open source projects.
Keep the list of sentimentized emoji around, so we can search for those
later.

Signed-off-by: Sarah Sharp <sharp@otter.technology>